Separate some SortedSet types into new files
